Mock functions moved to mock module
